+++ site/com_joomgallery/language/en-GB/com_joomgallery.ini
@@ -58,7 +58,6 @@ COM_JOOMGALLERY_ACTIONS="Actions"
 COM_JOOMGALLERY_LOAD_MORE="Load More"
 COM_JOOMGALLERY_NO_MORE_IMAGES="No more images"
 
-
 ;Types
 COM_JOOMGALLERY_THUMBNAIL="Thumbnail"
 COM_JOOMGALLERY_ORIGINAL="Original"
@@ -123,10 +122,12 @@ COM_JOOMGALLERY_ITEM_STATE_ERROR="State change failed with the following error: 
 COM_JOOMGALLERY_ITEM_SAVE_SUCCESSFUL="Item successfully saved."
 COM_JOOMGALLERY_NOTE_DEVELOPMENT_VERSION="Attention!<br />----------------<br />This version of JoomGallery is still under development. Do not use this version on a live website. It is intended for testing purposes only..."
 
-
 ;Views
 COM_JOOMGALLERY_IMAGE_INFO="Image information"
 COM_JOOMGALLERY_IMAGE_NEW="New Image"
+COM_JOOMGALLERY_CATEGORY_PASSWORD_PROTECTED="This category is password protected"
+COM_JOOMGALLERY_CATEGORY_BUTTON_UNLOCK="Unlock"
+COM_JOOMGALLERY_CATEGORY_PASSWORD_INCORRECT="Incorrect password"
 
 ;Configuration
 COM_JOOMGALLERY_CONFIG_TYPEPARAMS_INFO="Current global parameter is calculated based on the Configuration Set of the current user.<br />The parameter applyed in the end may be different!"
@@ -138,7 +139,6 @@ COM_JOOMGALLERY_CONFIG_SHOW_GEOTAGGING="Show Map?"
 COM_JOOMGALLERY_CONFIG_SHOW_GEOTAGGING_IMG_LONG="Do you want to add a map showing where the image was taken?{tip}It is possible to use the GPS data of the original image to show as location in a Map. If the image contains valid GPS data within the Exif information an additional area with the map will be displayed in the detail view."
 COM_JOOMGALLERY_CONFIG_SHOW_IPTCDATA="Show IPTC data?"
 COM_JOOMGALLERY_CONFIG_SHOW_IPTCDATA_IMG_LONG="Shall the IPTC data of the image be shown if available?{tip}If \"Yes\" by the following options you can decide which information will be published. Otherwise they are ineffective."
-
 COM_JOOMGALLERY_CONFIG_OFFER_DOWNLOAD="Offer Downloads"
 COM_JOOMGALLERY_CONFIG_OFFER_DOWNLOAD_CAT_LONG="Here you can set whether the download is allowed in this category.{tip}This can overwrite the setting in tab 'User Access Rights->Download->Offer Downloads' in JoomGallery Configuration"
 COM_JOOMGALLERY_CONFIG_ALLOW_COMMENT="Allow Comments"
